Allowing TMUX pane titles to persist
When running within tmux, attempt the following at the prompt
    
    ```bash
    > tmux set -g pane-border-status bottom
    > tmux set -g allow-rename off
    > tmux set -g pane-border-format "< #T >"
    > tmux split-window
    
    > tmux select-pane -T "New Title"
    ```    
Prior to this change, the pureline_pre function is constantly setting the gui window title.  This is the same value that TMUX uses to determine the pane title.  This 'resetting' of the xterm title overwrites the tmux pane title.  For this reason, and only within properly branded TMUX sessions, we do not reset the window title.