Fixes when inherited PROMPT_COMMAND is not empty
1. [FIX] function 'pureline_pre' now returns the saved error code of the last executed command.
=> Does not break an inherited PROMPT_COMMAND that requires that value.

2. [FIX] Defines PROMPT_COMMAND as a call to a single function,
which is dynamically generated using 'eval' in a robust way, so as to work whatever the content of the inherited PROMPT_COMMAND.
=> No need to strip trailing whitespaces
=> No need to check/add a trailing semicolon
=> No need to handle any other (undiscovered) special case

Note: the "new" function is called 'pureline_ps1', while the "old" one is renamed 'pureline_post'.

3. [IMPROVEMENT] Defining PROMPT_COMMAND as a single function call simplifies a lot the integration of pureline with other prompt-modifying tools:
for example, the 'shell integration' feature in VSCode integrated terminals does not break anymore when using pureline.